requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.7
appdirs==1.4.4
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0
bcrypt==3.2.0; python_version >= '3.6'
blinker==1.4
bpyutils==0.4.0
brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.2
coverage==6.3.2
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dnspython==2.2.1; python_version >= '3.6' and python_version < '4'
dominate==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.3
faker==13.3.5
flask-admin==1.6.0
flask-assets==2.0
flask-bcrypt==1.0.1
flask-bootstrap==3.3.7.1
flask-compress==1.11
flask-debugtoolbar==0.13.1
flask-login==0.6.0
flask-mail==0.9.1
flask-migrate==3.1.0
flask-rq==0.2
flask-script==2.0.6
flask-sqlalchemy==2.5.1
flask-sslify==0.1.5
flask-testing==0.8.1
flask-wtf==1.0.1
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
honcho==1.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1; python_version < '3.9'
itsdangerous==2.1.2
jinja2==3.1.1
jsmin==3.0.1
jsonpickle==2.1.0
mako==1.2.0
markupsafe==2.1.1
packaging==21.3; python_version >= '3.6'
pipupgrade==1.12.0
psycopg2-binary==2.9.3
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyodbc==4.0.32
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
raygun4py==4.3.2
redis==4.2.2
requests==2.27.1
rq==1.10.1
six==1.16.0
sqlalchemy==1.4.35
text-unidecode==1.3
typing-extensions==4.2.0; python_version < '3.8'
urllib3==1.26.9
visitor==0.1.3
webassets==2.0
wrapt==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wtforms
zipp==3.8.0; python_version < '3.10'
pycep_correios==5.0.0
geopy==2.2.0
pandas==1.1.3
unidecode==1.3.4
xlrd
openpyxl
Werkzeug <= 2.1.2
flask == 2.1.2
flask-restx >= 0.5.1